Some changes to compile without errors.

+++ order-mgr/src/main/java/de/dlr/proseo/ordermgr/rest/model/RestUtil.java
@@ -110,7 +110,7 @@ public class RestUtil {
 			}
 			
 			if (modelJobStep.getPriority() != null) {
-				pjs.setPriority(modelJobStep.getPriority());
+				pjs.setPriority(modelJobStep.getPriority().longValue());
 			}
 
 			pjs.setVersion((long) modelJobStep.getVersion());
